Made some updates to the front end
The main.py page in allows us to run the app straight.

It calls the init.py which is the controller of sorts from there we just have to import wtv pages we want to use and itll run for us too.

Updated the base template to better reflect the pages.